research/requirements.txt
asyncpg==0.30.0
contourpy==1.3.1
cycler==0.12.1
dnspython==2.7.0
fonttools==4.55.6
joblib==1.4.2
kiwisolver==1.4.8
matplotlib==3.10.0
motor==3.6.1
numpy==2.2.2
packaging==24.2
pillow==11.1.0
psycopg2-binary==2.9.10
pymongo==4.9.2
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2024.2
six==1.17.0
tqdm==4.67.1
tzdata==2025.1
